Configure country currency master data

+++ ginum-backend-main/src/main/java/com/example/GinumApps/repository/CountryRepository.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 
 
 @Repository
-public interface CountryRepository extends JpaRepository<Country, Long> {
+public interface CountryRepository extends JpaRepository<Country, Integer> {
     Optional<Country> findByName(@NotNull String name);
 }
